fix: persist 적용 안되는 부분 수정
